changed default 'current-image' to rumu
TODO: the default 'current-image' should be handled solely by the first element in the list of thumbnails, as this causes a discrepency with the first thumbnail and the current image if both aren't updated accordingly.

'current-image' also seems to be the first image loaded, which is what is used as the default image for embedding (if the og/twitter meta tags aren't used), so it's questionable whether leaving it empty would break default embedding.